nvim: Fix width of REPL window
I noted in ad19eaa that this shouldn't be necessary in nvim 0.11, since
it reflows text in terminal buffers. However, the text is still
truncated when it gets too narrow (maybe due to text moving into the
scrollback?), such as when the the window is resized to half its width
while the REPL is in the right half, in which case the REPL window is
briefly resized to the minimum width.