nvim: Make indentation heuristic more robust
The previous implementation could give incorrect results when just a
single line is incorrectly indented. Now both whether to set expandtab
and the value of tabstop are determined by the most common occurrence in
the buffer (the latter uses the difference between lines rather than the
absolute indentation, like vim-sleuth, since otherwise the majority vote
would be incorrect).

This no longer handles the case where mixed tabs and spaces are used in
the same line, since that style is uncommon.

The logic is also extracted to a user command, so that the indentation
settings can be manually detected from a buffer populated after initial
read.