nvim: Improve indentation detection
First, when nothing can be learned from the buffer contents, it keeps
the existing settings for expandtab and shiftwidth. This makes no
difference on BufReadPost, but can be relevant when running
:DetectIndent manually.

Second, when any line contains mixed tabs and spaces, use traditional
tabstop=8 for maximum compatibility. These lines are ignored when
calculating shiftwidth, since the spaces could be for alignment rather
than indentation.