nvim: Disable vimtex's completion of closing braces
I had initially enabled it so that I don't need to type the closing
brace after the "begin environment" mapping. It caused problems with
multiple citations in one command, which I attempted to fix in 2bf6437,
but that results in unituitive behavior when a citation should be
followed by a comma.