bin: Fix rounding issue in brightness script
When changing by fixed increments of N, the brightness would not stay a
multiple of N, due to always rounding down when converting between the
absolute value and a percentage. This fixes it by rounding down in one
direction and up the other.

It no longer sets the minimum brightness to 1.