kmonad: Remap copilot key to match other keys right of ralt
The key sends lmet+lsft+f23, with the presses in that order (although I
suspect the order could be different on other devices). The lmet key
switches to the fn layer, where, fortunately, lsft is mapped to XX. So
effectively, the copilot key presses f23 on the fn layer, which is why
that key is changed from ret to _ (which falls through to ret_fn).