nvidia: Fix issue where GPU would not suspend
Before this change, if the X server was ever not running while an
external display is connected (via the Nvidia GPU), then the GPU would
no longer suspend, requiring a reboot to fix.

Arch sets modeset=1 and fbdev=1 for the nvidia_drm kernel module by
default. Resetting either fixes the issue; we choose the latter, since
the former is required for Wayland. (For all the Wayland compositors I
tried, it was necessary to unset the __EGL_VENDOR_LIBRARY_FILENAMES
override from ~/.zshenv for the compositor to use the external monitor.)