ipython: Pad images to multiple of cell size
This prevents inconsistent scaling from one image to the next. In fact,
if cell_width and cell_height are correct, it prevents any scaling at
all. This was particularly noticeable with LaTeX displayed by SymPy, for
example.

This adds Pillow as a dependency, although it is already a dependency of
Matplotlib. Another advantage of this is that we are no longer
restricted to PNG. I've only added JPEG for now, since that is the only
other format supported by Pillow that is mentioned in the IPython code.