zsh: Remove some symbols from prompt
The Powerline separator doesn't work well when color information is
lost, such as when using kitty's hints mode.

Ideally we could use the COLORTERM variable to decide whether to use the
RGB variant of the prompt, but this variable isn't propagated over SSH
without configuration on both client and server, so instead assume
everything except the Linux TTY supports it.