kitty: Configure key mappings
There are no mappings to manage split windows and tabs, since I prefer
using the window manager for that. This is the primary reason why the
default mappings are cleared and configured from scratch, but that has
also given the opportunity to choose mappings that are easier to
remember.

Most of these mappings use a verb-object model. Some of the combinations
seem unlikely to be useful, but are included for completeness.

The last_visited_cmd_output object could be replaced by
first_cmd_output_on_screen, since they have the same behavior directly
after jumping to a prompt with ctrl+shift+[/], and the latter has more
straightforward behavior before jumping or after scrolling away from the
prompt. However, there is not much point in using the pager verb on a
recent command that fits entirely on the screen, and one can jump to a
command or select it with the mouse, then use last_visited_cmd_output to
refer back to it multiple times.

The mouse mapping to select a command's output is set to
ctrl+shift+alt+right, for similarity with the default ctrl+shift+right
mapping to show the output in the pager, and the ctrl+shift+alt drag and
triple-click mappings for non-standard selections.

I don't see any difference between 'scroll_prompt_to_top' and
'clear_terminal scroll active' - both act as a better ctrl+l that
preserves all output in the scrollback buffer. I chose the former
(mapped to ctrl+shift+z>t) for symmetry with 'scroll_prompt_to_bottom',
which is useful because kitty does not by default scroll lines back onto
the screen when the window height is increased. That behavior can be
changed with 'scrollback_fill_enlarged_window yes', but I left it as is
for now because with other terminals I had found that annoying after
clearing the screen. I considered mapping ctrl+shift+l (or even just
ctrl+l) to 'combine : clear_terminal scroll active : send_key ctrl+l',
which would perform this better clear in shells or REPLs running
directly in kitty, or a legacy clear when running in the nvim terminal,
for example. An alternative would be to update the terminfo description
so that programs use this escape sequence automatically. One can run
    tic -x - <<<'xterm-kitty|KovIdTTY,clear=\r\e[J\e[H\e[22J,use=xterm-kitty'
to create an updated description in ~/.terminfo, and set 'terminfo none'
so kitty doesn't override it with its own. However, not all REPLs
(e.g. ipython) respect this, and it doesn't work for zsh (which does),
because of my two-line prompt. The kitty documentation shows how ctrl+l
can be mapped in zsh to print the escape sequence and redraw the
prompt, but the sequence must be changed to '\e[F\e[J\e[H\e[22J'.