kitty: Correctly mark start of command when ran as xterm
kitty does not use -e to mark remaining arguments as the command to run;
-- should be used instead. As an undocumented feature, it seems to
ignore -e for compatibility.

We also pass -e from the xmonad configuration, but that is intended to
be somewhat agnostic of the terminal, and the next argument is sh, which
is not at risk of being confused for an option.