kitty, zsh: Change TERM for better compatiblity over ssh
kitty defaults to using xterm-kitty and provides its own terminfo entry.
However, an entry called kitty has been distributed with ncurses since
2018, which should by now be available on most servers. If necessary,
the variable can be overridden per server in ~/.ssh/config.

An alternative would be to set terminfo_type to direct (or
TERMINFO=$(infocmp -qx0Q2) for the same effect), and configure both the
ssh client and server to send the TERMINFO variable. Or use the ssh
kitten, which copies the entry to ~/.terminfo on the server.