sb: Fix mounting paths under sb-share
The add-mount function uses realpath -s, which doesn't follow symlinks,
so that when you specify a path containing a symlink, the same path is
available in the sandbox (but it's mounted there, so it's no longer a
symlink). However, this breaks when the symlink already exists in the
sandbox, such as when mounting a file under ~/sb-share. Instead of
attempting to track all such symlinks, this just checks for the
~/sb-share case. We could also ignore attempts to mount files under that
path, since they are already available in the sandbox, but that would
silently make paths writable when a read-only mount is requested.